+++ packages/eas-cli/src/graphql/mutations/PostHogMutation.ts
@@ -5,19 +5,19 @@ import { ExpoGraphqlClient } from '../../commandUtils/context/contextUtils/creat
 import { withErrorHandlingAsync } from '../client';
 import { CreatePostHogAccountRequestInput, SetupPostHogProjectInput } from '../generated';
 import {
-  PostHogOrganizationConnectionData,
   PostHogOrganizationConnectionFragmentNode,
   PostHogProjectData,
   PostHogProjectFragmentNode,
+  StartPostHogConnectionResult,
 } from '../types/PostHogConnection';
 
-type CreatePostHogAccountRequestMutation = {
+type StartPostHogConnectionMutation = {
   posthogOrganizationConnection: {
-    createPostHogAccountRequest: PostHogOrganizationConnectionData;
+    startPostHogConnection: StartPostHogConnectionResult;
   };
 };
 
-type CreatePostHogAccountRequestMutationVariables = {
+type StartPostHogConnectionMutationVariables = {
   input: CreatePostHogAccountRequestInput;
 };
 
@@ -42,32 +42,36 @@ type DeletePostHogProjectMutationVariables = {
 };
 
 export const PostHogMutation = {
-  async createPostHogAccountRequestAsync(
+  async startPostHogConnectionAsync(
     graphqlClient: ExpoGraphqlClient,
     input: CreatePostHogAccountRequestInput
-  ): Promise<PostHogOrganizationConnectionData> {
+  ): Promise<StartPostHogConnectionResult> {
     const data = await withErrorHandlingAsync(
       graphqlClient
-        .mutation<
-          CreatePostHogAccountRequestMutation,
-          CreatePostHogAccountRequestMutationVariables
-        >(
+        .mutation<StartPostHogConnectionMutation, StartPostHogConnectionMutationVariables>(
           gql`
-            mutation CreatePostHogAccountRequest($input: CreatePostHogAccountRequestInput!) {
+            mutation StartPostHogConnection($input: CreatePostHogAccountRequestInput!) {
               posthogOrganizationConnection {
-                createPostHogAccountRequest(input: $input) {
-                  id
-                  ...PostHogOrganizationConnectionFragment
+                startPostHogConnection(input: $input) {
+                  __typename
+                  ... on PostHogOrganizationConnection {
+                    id
+                    ...PostHogOrganizationConnectionFragment
+                  }
+                  ... on PostHogPendingConnection {
+                    url
+                  }
                 }
               }
             }
             ${print(PostHogOrganizationConnectionFragmentNode)}
           `,
-          { input }
+          { input },
+          { additionalTypenames: ['PostHogOrganizationConnection'] }
         )
         .toPromise()
     );
-    return data.posthogOrganizationConnection.createPostHogAccountRequest;
+    return data.posthogOrganizationConnection.startPostHogConnection;
   },
 
   async setupPostHogProjectAsync(